Restructure frontend added dash nav
