+++ app/javascript/components/App.jsx
@@ -5,11 +5,11 @@ import Home from '../components/shared/home';
 import About from '../components/shared/about';
 import Contact from '../components/shared/contact';
 import Nav from '../components/shared/nav';
+import DashNav from '../components/private/dash_nav';
 import Login from './login';
 import Signup from './signup';
 
 require('../style/App.scss');
-
 
 class App extends React.Component {
   constructor(){
@@ -36,9 +36,59 @@ class App extends React.Component {
     } else {
       this.setState({ current_user: null })
     }
+    this.state.current_user ? this.getUserData() : console.log("skiped")
   }
   tokenHandler=()=>{
     return this.state.token
+  }
+  dataParser =(data)=>{
+    const subj = Object.keys(data)[0];
+    const detail = data[subj];
+    if(Array.isArray(detail)){
+      Object.values(data)[0].map((k, v) =>{
+        k.amount = parseInt(k.amount)
+      });
+      this.setState(data)
+    } else {
+      data[subj].amount = parseInt(data[subj].amount)
+      this.setState(data)
+    }
+  }
+  fetchUserData = (endPoint, method, data) => {
+    const baseUrl = 'http://localhost:3000/api/v1';
+    const param = { 
+      method: method,
+      headers: {
+        "Content-Type": "application/json",
+        "charset": "UTF-8",
+        "Authorization": "",
+        "mode": "cors"
+      }
+    }
+    method == "POST"|"EDIT"|"DELETE" ? param["body"] = JSON.stringify(data) : ""
+    const token = sessionStorage.getItem("token");
+    if (this.state.current_user) {
+      param.headers["Authorization"] = token;
+      fetch( baseUrl +"/"+ endPoint, param)
+      .then(res => res.json())
+      .then ((result) => {
+        param.body ? this.showMessage(result) : this.dataParser(result)
+      }).catch(
+        e =>{ console.log('Asset errors',e);}
+      );
+    } else {
+      console.log("No Token so far")
+    }     
+  }
+  getUserData() {
+    ["assets","commitments","networth"].forEach(
+      (item) => { this.fetchUserData(item, "GET")})
+  }
+  componentDidUpdate(prevState, prevProps) {
+    if (this.state.current_user !== prevProps.current_user) {
+      this.getUserData();
+    }
+    console.log(this.state.current_user,prevProps.current_user)
   } 
   render(){
     const currentUser       = this.state.current_user;
@@ -46,12 +96,16 @@ class App extends React.Component {
     const updateCurrentUser = this.updateCurrentUser;
     const toggleLogin       = this.toggleLogin;
     const fetchUserData     = this.fetchUserData;
-    const data              = this.state;
     const getUserData       = this.getUserData;
+    const data = {
+      asset: this.state.asset,
+      commit: this.state.commits,
+      net:  this.state.net
+    };
     return (      
       <BrowserRouter>
         <div className="content" > 
-          <Nav />                           
+          {currentUser ? <DashNav/> : <Nav />}                           
             <div className="route">
               <Switch>
                 <Route exact path="/" render= { (props)=> <Home { ...props}/>}/>
@@ -68,7 +122,7 @@ class App extends React.Component {
                     getUserData = { getUserData }
                     data = { data }
                     />
-                  } 
+                  }
                 />
               </Switch>
             </div>    
@@ -79,4 +133,4 @@ class App extends React.Component {
   }
 }
 
-export default App;
+export default App;
