+++ app/javascript/components/private/dash_board.jsx
@@ -1,13 +1,11 @@
-import React from 'react'
+import React from 'react';
 import { Route, Link, Switch } from "react-router-dom";
-import Assets from './assets'
-import Commitments from './commitments'
-import Networth from './networth'
-import Logout from './logout'
+import Assets from './assets';
+import Commitments from './commitments';
+import Networth from './networth';
+import Logout from './logout';
 import Login from '../login';
 import Signup from '../signup';
-import Section from './section';
-import Form from './form'
 
 class DashBoard extends React.Component {
   constructor(props) {
@@ -25,45 +23,7 @@ class DashBoard extends React.Component {
     console.log(result.message)
     this.setState({message: result.message})
   }
-  dataParser =(data)=>{
-    const subj = Object.keys(data)[0];
-    const detail = data[subj];
-    if(Array.isArray(detail)){
-      Object.values(data)[0].map((k, v) =>{
-        k.amount = parseInt(k.amount)
-      });
-      this.setState(data)
-    } else {
-      data[subj].amount = parseInt(data[subj].amount)
-      this.setState(data)
-    }
-  }
-  fetchUserData = (endPoint, method, data) => {
-    const baseUrl = 'http://localhost:3000/api/v1';
-    const param = { 
-      method: method,
-      headers: {
-        "Content-Type": "application/json",
-        "charset": "UTF-8",
-        "Authorization": "",
-        "mode": "cors"
-      }
-    }
-    method == "POST"|"EDIT"|"DELETE" ? param["body"] = JSON.stringify(data) : ""
-    const token = sessionStorage.getItem("token");
-    if (this.props.currentUser) {
-      param.headers["Authorization"] = token;
-      fetch( baseUrl +"/"+ endPoint, param)
-      .then(res => res.json())
-      .then ((result) => {
-        param.body ? this.showMessage(result) : this.dataParser(result)
-      }).catch(
-        e =>{ console.log('Asset errors',e);}
-      );
-    } else {
-      console.log("No Token so far")
-    }     
-  }
+  
   sumData = (value) =>{
     const reducer = (accumulator, currentValue) =>
      accumulator + currentValue;
@@ -113,22 +73,12 @@ class DashBoard extends React.Component {
     this.setState({formOpen: !this.state.formOpen});
   }
   componentDidMount(){
-    this.props.currentUser ? this.getUserData() : console.log("skiped")
-  }
-  componentDidUpdate(prevProps) {
-    if (this.props.currentUser !== prevProps.currentUser) {
-      this.getUserData();
-    }
-  }
-  getUserData() {
-    ["assets","commitments","networth"].forEach(
-      (item) => { this.fetchUserData(item, "GET")})
+    //this.props.currentUser ? this.props.getUserData() : console.log("skiped")
   }
   render() {
-    const fetchUserData = this.fetchUserData;
+    const fetchUserData = this.props.fetchUserData;
     const submitHandler = this.submitHandler;
-    const data = this.state;
-    const token = this.getUserToken;
+    const data = this.props.data;
     const currentUser= this.props.currentUser;
     const toggleLogin = this.props.toggleLogin;
     const updateCurrentUser = this.props.updateCurrentUser;
@@ -146,12 +96,7 @@ class DashBoard extends React.Component {
               toggleLogin={ toggleLogin }
               updateCurrentUser = { updateCurrentUser }
               user={ currentUser }
-            />
-            <div className="user_links">              
-              <Link to={ "/assets" }>Assets</Link>              
-              <Link to={ "/commits" }>Commitments</Link>
-              <Link to={ "/networth" }>Net Worth</Link>              
-            </div>            
+            />          
             <Switch>
               <Route exact path="/assets" 
                 render={
@@ -169,7 +114,7 @@ class DashBoard extends React.Component {
               <Route  path="/commits" 
                 render={
                   (props)=> <Commitments { ...props}
-                    data={ data.commits }
+                    data={ data.commit }
                     fetchUserData={ fetchUserData }
                     token={this.getUserToken}
                     currentUser={currentUser}
