+++ app/views/site/landing.html.erb
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <!-- <link rel="icon" href="data:;base64,=">  added to stop req for favicon--> -->
+    <!-- <link rel="icon" href="data:;base64,=">-->
     <%= favicon_link_tag 'public/favicon.ico' %>
     <%= javascript_pack_tag 'application'%>
     <%= stylesheet_pack_tag 'application'%>
